+++ example/lib/profile/data/model/api_user_response.dart
@@ -6,10 +6,9 @@ part 'api_user_response.freezed.dart';
 part 'api_user_response.g.dart';
 
 @freezed
-class ApiUserResponse with _$ApiUserResponse {
-  const factory ApiUserResponse({
-    required List<User> results,
-  }) = _ApiUserResponse;
+sealed class ApiUserResponse with _$ApiUserResponse {
+  const factory ApiUserResponse({required List<User> results}) =
+      _ApiUserResponse;
 
   factory ApiUserResponse.fromJson(Map<String, dynamic> json) =>
       _$ApiUserResponseFromJson(json);