+++ example/lib/profile/data/model/street.dart
@@ -4,11 +4,8 @@ part 'street.freezed.dart';
 part 'street.g.dart';
 
 @freezed
-class Street with _$Street {
-  const factory Street({
-    required int number,
-    required String name,
-  }) = _Street;
+sealed class Street with _$Street {
+  const factory Street({required int number, required String name}) = _Street;
 
   factory Street.fromJson(Map<String, dynamic> json) => _$StreetFromJson(json);
 }