+++ example/lib/profile/presentation/cubit/user_state.dart
@@ -1,8 +1,6 @@
 part of 'user_cubit.dart';
 
 @freezed
-class UserState with _$UserState {
-  const factory UserState({
-    User? user,
-  }) = _UserState;
+sealed class UserState with _$UserState {
+  const factory UserState({User? user}) = _UserState;
 }