+++ lib/chopper/datetime_converter.dart
@@ -1,10 +1,10 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:intl/intl.dart';
+import 'package:json_annotation/json_annotation.dart';
 
 enum _DateFormats {
   yyyyMMdd,
-// Name contains the exact pattern for date formatting.
-//ignore: constant_identifier_names
+  // Name contains the exact pattern for date formatting.
+  //ignore: constant_identifier_names
   HHmm,
 }