+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog
+
+## 0.0.12
+* Updated dependencies
+* Added test_util: catch presentation event helper, DeviceSizes
+* Added typedef for single bool methods.
+* Updated formatting
+* exported more files
+
 ## 0.0.11
 Added optional secondary, secondaryFill, tertiary and tertiaryFill ColorGroups to the Kleurplaat